refactor(logic): break orders->ai feedstock-extraction edge (Refs #3290)
Relocate the feedstock-extraction resource-id closure
(feedstockExtractionResourceIdsForPlayer and its transitive helpers) from
the Full-AI civilian-work-selection library (lib/src/ai/) into the orders
domain (lib/src/orders/feedstock_extraction_targets.dart).

This eliminates the orders -> ai import edge: order_suggestion_work no
longer imports lib/src/ai/, and the AI selection library now depends on the
orders-owned extraction gates, matching the one-way DAG the epic mandates
(colonizethis_ai_contracts depends on colonizethis_orders, not the reverse).

Behavior-preserving intra-package move: function bodies are identical, three
shared helpers (regimentCountForPlayer, isBelowQuotaZeroNwSeller,
sellerImprovementInputFeedstockResourceIds) are promoted from private to
public so the kept AI-only acquisition/staging code can still consume them,
and ai_api.dart re-exports the moved symbols from their new orders location
so colonizethis_ai's narrow contract is unchanged.
